requirements.txt
pytorch-lightning
flax
transformers
omegaconf==2.0.0
wandb